requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isort==5.13.2
jinja2==3.1.4
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mypy==1.10.1
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
rich==13.7.1
ruff==0.5.3
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.32
sqlalchemy-utils==0.41.2
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
